actix: split out integration into separate crate
