+++ testing/Cargo.toml
@@ -6,17 +6,19 @@ workspace = ".."
 edition = "2018"
 
 [features]
-actix = ["actix-web", "bytes", "askama/with-actix-web"]
+actix = ["actix-web", "bytes", "askama_actix"]
 default = []
 full = ["actix", "with-iron", "serde-json", "with-gotham"]
 serde-json = ["serde_json", "askama/serde-json"]
 with-rocket = ["rocket", "askama/with-rocket"]
 with-iron = ["iron", "askama/with-iron"]
-with-gotham = ["gotham", "askama/with-gotham", "mime", "hyper"]
+with-gotham = ["gotham", "askama_gotham", "mime", "hyper"]
 
 [dependencies]
 actix-web = { version = "0.7", optional = true }
 askama = { path = "../askama", version = "*" }
+askama_actix = { path = "../askama_actix", version = "*", optional = true }
+askama_gotham = { path = "../askama_gotham", version = "*", optional = true }
 bytes = { version = "0.4", optional = true }
 iron = { version = "0.6", optional = true }
 rocket = { version = "0.4", optional = true }
